refactor: 검색 관련 쿼리 구조 변경 및 새로운 쿼리 파일 추가
