+++ apps/client/src/shared/hooks/use-related-search.ts
@@ -1,9 +1,7 @@
 import { useQueries } from '@tanstack/react-query';
 
-import {
-  SEARCH_ARTIST_QUERY_OPTION,
-  SEARCH_PERFORMANCE_QUERY_OPTION,
-} from '@shared/apis/search/search-queries';
+import { SEARCH_ARTIST_QUERY_OPTIONS } from '@shared/apis/search/search-artist-queries';
+import { SEARCH_PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/search/search-performance-queries';
 
 interface KeywordProps {
   keyword: string;
@@ -13,8 +11,8 @@ interface KeywordProps {
 export const useRelatedSearch = ({ keyword, enabled }: KeywordProps) => {
   return useQueries({
     queries: [
-      SEARCH_ARTIST_QUERY_OPTION.SEARCH_RELATED_KEYWORD(keyword, enabled),
-      SEARCH_PERFORMANCE_QUERY_OPTION.SEARCH_RELATED_PERFORMANCES(
+      SEARCH_ARTIST_QUERY_OPTIONS.SEARCH_RELATED_KEYWORD(keyword, enabled),
+      SEARCH_PERFORMANCE_QUERY_OPTIONS.SEARCH_RELATED_PERFORMANCES(
         keyword,
         enabled,
       ),
